+++ archivos/bodega.html
@@ -13,14 +13,29 @@
 </head>
 
 <body>
-
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
         <div class="container-fluid sectionNav">
-            <div class="w-25  p-3" style="background-color: #ecf19e;">
-                <div class="iconoMenuFijo">
-                    <img src="../multimedia/logo02.png" alt="iconoMenu" width="140px" height="70px">
-                </div>
-            </div>
+            <ul class="liCambio navbar-nav">
+                <li>
+                    <div class="w-25  p-3" style="background-color: #ecf19e;">
+                        <div class="iconoMenuFijo">
+                            <img src="../multimedia/logo02.png" alt="iconoMenu" width="140px" height="70px">
+                        </div>
+                    </div>
+                </li>
+                <li>
+                    <div>
+                        <!--es para la seccion de noticias e interes-->
+                        <button class="btn btn-primary classButton " id="posiProduccion" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasRight" aria-controls="offcanvasRight">Interes</button>
+
+                        <!--para el navegador lateral-->
+                        <button class="btn btn-primary classButton" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasExample" aria-controls="offcanvasExample">
+                            Sub Menu
+                        </button>
+                    </div>
+
+                </li>
+            </ul>
             <div class="w-50 p-3 " style="background-color: #ecf19e;">
                 <a class="nav-link active" aria-current="page" href="../index.html">
                     <h1>PANISOF</h1>
@@ -72,8 +87,6 @@
         <div class="row">
             <div class="col-2 col-md-2">
                 <!--es para la seccion de noticias e interes-->
-                <button class="btn btn-primary classButton" id="posiProduccion" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasRight" aria-controls="offcanvasRight">Interes</button>
-
                 <div class="offcanvas offcanvas-end" style="background-color: #ecf19e;" tabindex="-1" id="offcanvasRight" aria-labelledby="offcanvasRightLabel">
                     <div class="offcanvas-header">
                         <h5 id="offcanvasRightLabel">Interes</h5>
@@ -93,10 +106,6 @@
                     </div>
                 </div>
                 <!--para el navegador lateral-->
-                <button class="btn btn-primary classButton" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasExample" aria-controls="offcanvasExample">
-                     Sub Menu
-                </button>
-
                 <div class="offcanvas offcanvas-start" style="background-color: #ecf19e;" tabindex="-1" id="offcanvasExample" aria-labelledby="offcanvasExampleLabel">
                     <div class="offcanvas-header">
                         <h5 class="offcanvas-title" id="offcanvasExampleLabel">Seleccione</h5>
@@ -105,7 +114,7 @@
                     <div class="offcanvas-body" style="background-color: #ecf19e;">
                         <div class="dropdown mt-3">
                             <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-bs-toggle="dropdown">
-                                Sub menu bodega
+                                Sub menu
                             </button>
                             <ul class="dropdown-menu fontLi" aria-labelledby="dropdownMenuButton">
                                 <li><a class="dropdown-item fontLi" href="#proveeBodega">Proveedor</a></li>
