+++ archivos/workspace.code-workspace
@@ -1,0 +1,16 @@
+{
+    "folders": [
+		{
+			"path": ".."
+		},
+		{
+			"path": "D:\\entregas bootcamp\\modulo1 botcamp\\Actividad4U2"
+		},
+		{
+			"path": "D:\\entregas UPB\\programacion\\java ciclo2"
+		}
+	],
+    "settings": {
+        "liveServer.settings.multiRootWorkspaceName": "proyectoFinalIbero"
+    }
+}
